+++ scripts/index.js
@@ -20,10 +20,10 @@ function chMemoOnChange(e,cid){
 newTab("tab_links","リンク","");
 (function(){
     //Generates list of links in links tab
-    var content = '<table><tr><td><input type="text" id="link-addLink" placeholder="新しいリンク"></td><td><button id="link-btnLinkAdd" onclick="pso2c.newLink($(\'#link-addLink\').val(),null,()=>{location.reload();});">追加</button></td></tr>';
+    var content = '<table><tr><td><input type="text" id="link-addLink" placeholder="新しいリンク"></td><td><button id="link-btnLinkAdd" onclick="pso2c.newLink($(\'#link-addLink\').val(),null,()=>{refreshPage();});">追加</button></td></tr>';
     var links = pso2c.getLinks((r)=>{
         r.forEach((e) => {
-            content += `<tr><td><a href="#" onclick="electron.shell.openExternal('${e.link}')">${e.title}</a></td><td><button onclick="if(confirm('削除されます？')){pso2c.deleteLink(${e.pklink},()=>{location.reload();});}">削除</button></td></tr>`;
+            content += `<tr><td><a href="#" onclick="electron.shell.openExternal('${e.link}')">${e.title}</a></td><td><button onclick="if(confirm('削除されます？')){pso2c.deleteLink(${e.pklink},()=>{refreshPage();});}">削除</button></td></tr>`;
         });
         content += '</table>';
         $("#tab_links").append(content);
@@ -31,7 +31,7 @@ newTab("tab_links","リンク","");
 })();
 
 
-newTab("tab_list","キャラクターリスト","<button onclick=\"location.reload();\">もう一度読む</button><br><br><table id=\"tab_list_chList\"></table>");
+newTab("tab_list","キャラクターリスト","<br><br><table id=\"tab_list_chList\"></table>");
 $("#tab_list_chList").append(`<tr><td><input type="text" placeholder="キャラ名前" id="newChName"></td><td><button onclick="pso2c.addCh($('#newChName').val(),()=>{location.reload()})">追加</button></td><td>TAクールダウン状態</td><td>TAクールダウン完成時間</td></tr>`);
 pso2c.getChList((res)=>{
     res.forEach((r)=>{
@@ -64,7 +64,7 @@ pso2c.getChList((res)=>{
         var nowStr = nowDateStr + "T" + nowTimeStr;
 
         //Generating row for list page
-        var out = `<tr><td>`+CName+`</td><td><button onclick="if(confirm('削除されます？')){pso2c.deleteCh(`+cid+`,()=>{location.reload();});}">削除</button></td><td>`+(taCDDone?"<span style=\"color:green;font-size: 15pt;\">\u25CF</span>TA進行可能</td><td>":("<span style=\"color:red;font-size: 15pt;\">\u25CF</span>TAクールダウン中</td><td>"+taCDDate.toLocaleString()))+`</td></tr>`;
+        var out = `<tr><td>`+CName+`</td><td><button onclick="if(confirm('削除されます？')){pso2c.deleteCh(`+cid+`,()=>{refreshPage();});}">削除</button></td><td>`+(taCDDone?"<span style=\"color:green;font-size: 15pt;\">\u25CF</span>TA進行可能</td><td>":("<span style=\"color:red;font-size: 15pt;\">\u25CF</span>TAクールダウン中</td><td>"+taCDDate.toLocaleString()))+`</td></tr>`;
         $("#tab_list_chList").append(out);
 
         //Adding new tab for detailed character information
@@ -81,9 +81,9 @@ pso2c.getChList((res)=>{
         content += `前回のTA：` + taDate.toLocaleString() + `<br>`;
         content += `クールダウン状態：` + (taCDDone?"<span style=\"color:green;font-size: 15pt;\">\u25CF</span>クールダウン完了":("<span style=\"color:red;font-size: 15pt;\">\u25CF</span>クールダウン中 ("+taCDDate.toLocaleString()+"まで)")) + "<br>";
         content += `<br>TA時間設定：<br>`;
-        content += `<button onclick="pso2c.TArun(`+cid+`,null,()=>{location.reload();})">現在</button><br><br>`;
-        content += `<input type="datetime-local" id="taTime`+cid+`" value="`+nowStr+`"><button onclick="pso2c.TArun(`+cid+`,$('#taTime`+cid+`').val(),()=>{location.reload();});">この時間でTAやった</button><br><br>`;
-        content += `<input type="datetime-local" id="taTime`+cid+`c" value="`+nowStr+`"><button onclick="var d = new Date($('#taTime`+cid+`c').val());d.setHours(d.getHours()-166);pso2c.TArun(`+cid+`,d.toLocaleString(),()=>{location.reload();});">この時点でクールダウン完了</button><br><br>`;
+        content += `<button onclick="pso2c.TArun(`+cid+`,null,()=>{refreshPage();})">現在</button><br><br>`;
+        content += `<input type="datetime-local" id="taTime`+cid+`" value="`+nowStr+`"><button onclick="pso2c.TArun(`+cid+`,$('#taTime`+cid+`').val(),()=>{refreshPage();});">この時間でTAやった</button><br><br>`;
+        content += `<input type="datetime-local" id="taTime`+cid+`c" value="`+nowStr+`"><button onclick="var d = new Date($('#taTime`+cid+`c').val());d.setHours(d.getHours()-166);pso2c.TArun(`+cid+`,d.toLocaleString(),()=>{refreshPage();});">この時点でクールダウン完了</button><br><br>`;
 
         content += `</div>`;
 
@@ -99,3 +99,19 @@ pso2c.getChList((res)=>{
 })
 
 $(".footer").append('<span style="color:#CCC;">PSO2 Tools '+pso2c.version+'</span>');
+
+$(document).ready(()=>{
+    setTimeout(()=>{
+        var urlparam = new URLSearchParams(location.search.slice(1));
+        if(urlparam.has("target")){
+            var tgtbtn = $(".tablinks[x-target="+urlparam.get("target")+"]");
+            if(tgtbtn.length > 0){
+                tgtbtn[0].click();
+            }else{
+                $(".tablinks[default]")[0].click();
+            }
+        }else{
+            $(".tablinks[default]").click();
+        }
+    },500);
+});