+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### New feature
 - A links page to store useful website links, and it auto fetches the title of that web page.
 
+### Improved
+- After refresh the page, automatically opens the tab was active before refresh.
+
 # [0.0.2-alpha]
 (2017-08-29)
 ### New feature